requirements.txt
pycryptodome
pwinput
pytz
tqdm
PyYAML
pathlib
requests
pyfiglet
colorama
termcolor